feat: Added functionality search errors
Not found components now accepts props with text to display. If there is no URL match, 'Page not found' is displayed, when there is no matches for the city in the API (404) 'City not found, please try again.' text is displayed. In other cases, 'There was an error, please try again.' message is displayed.