fix: fixed error handling in ResultsPage.tsx
If response code is 200 page content is rendered, response has some other code NotFound component is rendered. If the code is 404, 'City not found, please try again' is passed to NotFound component. If response code is neither 200 nor 404, 'There was an error, please try again.' is passed to NotFound component.