feat: last searches component logic is finished
function call for saving search query is moved to fetchAll function. It improves consistency and makes checking for duplicates easier (There was an edge case when city name could be written with or without '-' between. In this cases both names would be added to last searches. This case is now eliminated). There is some styling needed, but logic is all finished.

function for getting last searches from localstorage is now created, to avoid logic in LastSearches.tsx component.